Issue #27 Add edit menu entry
